# CPLEXDirect performance improvements