… pyomo/solvers/tests/checks/test_CPLEXDirect.py
… pyomo/solvers/tests/checks/test_CPLEXPersistent.py