This PR contains a number of performance improvements to the `CPLEXDirect` class, and by extension, the `CPLEXPersistent` class.

# Notes

There are just a couple of functions (if not lines) that are resulting in very large bottlenecks. On the whole, things are quite performant as-is.

## Batching "transactions" with CPLEX

The `cplex` package's Linear Constraints and Variable interfaces allow for batched transactions. I think an appropriate design is to generate all the necessary data and add these objects as one call to the `solver_model`. I've also removed unnecessary transactions such as resetting variable bounds immediately after adding that variable with an obsolete bound.

## Querying results from CPLEX

Asking the `cplex` solution for *specific* variables' values is extremely slow due to the conversion on `cplex`'s end to lookup the variables you've asked for. It is orders of magnitude faster to get the full solution vector. Even worse, using the "specific variable interface" to get the full solution vector. If we must get specific variables (ie. `_load_vars()`) then their index should be used instead of their name.

## Expected performance vs LP solvers

It would be interesting to benchmark specifically writing the LP file vs the total time spent interfacing with the CPLEX library in isolation. I think it's a fallacy to suggest that the Direct interfaces *should* be faster because "there's no file IO" as suggested in GitHub issues and on SO. Once the repn has been generated, writing a string to disk takes no time at all. If anything I would imagine interfacing with a third-party library should be expected to take *more* time. To me, the gains are realised when using the Persistent interfaces to apply incremental resolves of a model. The LP solver has to regen the entire model whereas the Persistent interface only has to generate the changes.
Not to mention that a slight decline in performance is a small price to pay for access to the full advanced functionality of the `cplex` package for those who need it.

# Benchmark

- Ran the "benchmark-ish" case in #51 ten times for each solver:

| **Solver** | `master`                      | This branch |
| ------------- | ------------- | ------------- |
| **LP**     | 10.74s to 11.06s. Mean 10.84s | 10.94s to 11.27s. Mean 11.08s |
| **Direct** | 32.48s to 34.00s. _Mean 33.14s_ | 11.67s to 12.23s. _Mean 12.00s_ |

- There was still a speedup when I set `n_steps = 10` so it's not as though we're optimising for very large models only
- The total time is now almost wholly taken up by (a) model build and (b) `generate_standard_repn()` which both of these interfaces have to do (and is therefore out-of-scope here).
- There are still a few inefficiencies but I didn't bother digging / fixing to shave part-seconds on a large test case. For example, generators are slower than list comprehension.
	There are many places where generators are used where the iterable is sufficiently small to consume into memory (eg. `ComponentSet.update()`)
- Clearly the LP solver appears to for whatever reason gotten worse as well. I reran it as a control and I don't really have an explanation for that... Might need the reviewer's help in case I've somehow inadvertently impacted that solver.

### TODO

- [x] Unit tests

## Related

- #51
- #331
- #1160
- #1169

### Legal Acknowledgement

By contributing to this software project, I have read the [contribution guide](https://pyomo.readthedocs.io/en/stable/contribution_guide.html) and agree to the following terms and conditions for my contribution:

1. I agree my contributions are submitted under the BSD license.
2. I represent I am authorized to make the contributions and grant the license. If my employer has rights to intellectual property that includes these contributions, I represent that I have received permission to make contributions and grant the required license on behalf of that employer.