@ruaridhw 2020-05-10
> Overall I think this is really good. My biggest concern is around the licensing of `nullcontext`, but I think that is easy to get around.

Interesting! I never thought about there being an issue with copying CPython code given all of its standard libraries are used anyway. Happy to drop the contextmanager design, your alternative is much more explicit.

> One final question: your new tests all rely on mock. Are there at least a few tests of this functionality that actually engage the cplex solver (just so that our tests will catch if the underlying cplex argument API drifts)?

All of the tests engage the CPLEX solver. `mock` is only used to "spy" on the API calls to the CPLEX solver rather than mocking the solver object. The tests can therefore assert on the underlying arguments that Pyomo provided CPLEX in addition to ensuring the API call actually worked. This is illustrated by assertions such as:

```python
self.assertEqual(opt._solver_model.linear_constraints.get_num(), 1)
```
which would fail if `_solver_model` was not an actual `cplex.Cplex()` object.