@ruaridhw 2020-05-11
> Happy to drop the contextmanager design, your alternative is much more explicit.

I've updated accordingly.

> the rest of this MR maintains identical behaviour to the existing implementation

Just to add a caveat to this... The _user-facing_ behaviour is identical however we are now adding linear constraints to the CPLEX object _after_ quadratic and SOS constraints. I guess it's fine for this to change though since this is an internal implementation detail and I don't think CPLEX is actually impacted if this order is changed. The order of the constraints _within_ each interface (Linear, Quadratic, SOS) is still the same.