@michaelbynum 2020-05-21
I think the test failures are due to older cplex versions. It looks like `set_objective([0]*num_cols)` only works in newer versions of cplex....