@jsiirola 2020-05-21
> I think the test failures are due to older cplex versions. It looks like `set_objective([0]*num_cols)` only works in newer versions of cplex....

Per the CPLEX [documentation](https://www.ibm.com/support/knowledgecenter/SSSA5P_12.7.1/ilog.odms.cplex.help/refpythoncplex/html/cplex._internal._subinterfaces.ObjectiveInterface-class.html#set_quadratic):
> If the quadratic objective function is separable, the entries of the list must all be of type float

I looked in `lib/python3.5/site-packages/cplex/_internal/_subinterfaces.py`, and there is an explicit test against the `float` type.

Changing `[0] * num_vars` to `[0.] * num_vars` should fix things.