@ruaridhw 2020-05-21
Great spot, thanks both. I've added explicit conversion of the coeffs to `float` inside `_CplexExpr` as well.