@jsiirola 2020-05-21
The one failing test was due to a codecov upload problem.  Merging.